requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
boto3==1.36.21
botocore==1.36.21
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.5
django-cors-headers==4.7.0
django-environ==0.12.0
django-storages==1.14.4
djangorestframework==3.15.2
idna==3.10
jmespath==1.0.1
numpy==2.2.2
pandas==2.2.3
pillow==11.1.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
s3transfer==0.11.2
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
tzdata==2024.2
urllib3==2.3.0
whitenoise==6.9.0
